Rename set-lock cli command and add status response
* Rename "set-lock" to just "lock"
* Print a status hash in addtion to the exit code
